Make helper function to convert data & time
